requirements.txt
pan-os-python==1.7.1
pan-python==0.17.0
